Update prep for 4.0.0 beta (#407)
* Prep for 4.0.0 beta

Update NodeJS versions

* Update action versions

* Bump package.json version

* Ensure default node version is correct

* Update build actions

* Remove NodeJS 22 for now

* Remove last s390 reference
