+++ .docker/Dockerfile.alpine
@@ -1,4 +1,4 @@
-ARG NODE_VERSION=16
+ARG NODE_VERSION=20
 ARG OS=alpine
 
 #### Stage BASE ########################################################################################################
