+++ .docker/Dockerfile.debian
@@ -1,4 +1,4 @@
-ARG NODE_VERSION=16
+ARG NODE_VERSION=20
 ARG OS=buster-slim
 
 #### Stage BASE ########################################################################################################
