+++ docker-custom/docker-debian.sh
@@ -7,7 +7,7 @@ echo "#########################################################################"
 
 docker build --rm --no-cache \
     --build-arg ARCH=amd64 \
-    --build-arg NODE_VERSION=16 \
+    --build-arg NODE_VERSION=20 \
     --build-arg NODE_RED_VERSION=${NODE_RED_VERSION} \
     --build-arg OS=buster-slim \
     --build-arg BUILD_DATE="$(date +"%Y-%m-%dT%H:%M:%SZ")" \
