+++ docker-custom/package.json
@@ -1,6 +1,6 @@
 {
     "name": "node-red-docker",
-    "version": "3.1.3",
+    "version": "4.0.0-beta.0",
     "description": "Low-code programming for event-driven applications",
     "homepage": "http://nodered.org",
     "license": "Apache-2.0",
@@ -29,9 +29,9 @@
         }
     ],
     "dependencies": {
-        "node-red": "3.1.3"
+        "node-red": "4.0.0-beta.0"
     },
     "engines": {
-        "node": ">=14"
+        "node": ">=18"
     }
 }
